pyproject.toml
[project]
name = "neurovoice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.15"
dependencies = [
    "pyyaml>=6.0.2",
    "transformers==4.35.2",
    "tts==0.22.0",
    "torch==2.6.0",
    "pyzmq==26.4.0",
    "ipykernel==6.20.2",
    "soxr==0.3.7",
    "graphviz==0.20.3",
    "matplotlib==3.8.0",
    "onnx==1.16.2",
    "onnxruntime==1.16.3",
    "openpyxl==3.1.2",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "torchmetrics==1.7.0",
    "torchview==0.2.6",
    "tqdm==4.67.1",
    "typing-extensions==4.13.0",
    "openai-whisper>=20250625",
]
